Create FullBodyDetectorv3.py
Code is used to detect a human's upper, lower and full body, using Computer Vision. Haar Cascades were used to detect the human. In order to find the direction of movement, an Optical Flow was mapped onto the grayscale image of the detected human. Doing vector math and finding averages of multi-dimensional matrices, we arrive at a value of the general velocity of the moving object. This moving value is used to determine direction, positive being towards the camera and vice versa. There is a 1.5 unit buffer extending from -0.75 to 0.75 to account for the low resolution camera used at the time of creation. Higher Resolution cameras and cameras that are capable of recording more frames  per second will require a lower buffer so that they don't miss a target. The population count is pushed to a database that uses a mongodb framework, which is planned to be used to be the backend of a webapp and an android app. The Code from v1-v3 was written in VS Code in Python 3.7.3 64-Bit, while the android app is being made on Android Studio